requirements.txt
fastapi
uvicorn
streamlit
pandas
sqlapsycopg2-binary
SQLAlchemy
python-dotenv
imaplib2
pdfplumber
pytesseract
Pillow
openpyxl
python-docx
openai
mysql-connector-python
plotly
python-multipart
emails
psycopg2-binary
streamlit-audiorecorder
audioop-lts
gpt4all
SpeechRecognition
gTTS
pydub
soundfile
